+++ lib/common.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:vs_node_view/data/offset_extension.dart';
@@ -6,6 +8,16 @@ import 'package:vs_node_view/data/vs_node_data.dart';
 import 'package:vs_node_view/data/vs_node_data_provider.dart';
 
 typedef VSNodeDataBuilder = VSNodeData Function(Offset, VSOutputData?);
+
+const _chars = 'AaBbCcDdEeFfGgHhIiJjKkLlMmNnOoPpQqRrSsTtUuVvWwXxYyZz1234567890';
+Random _rnd = Random();
+
+String getRandomString(int length) => String.fromCharCodes(
+      Iterable.generate(
+        length,
+        (_) => _chars.codeUnitAt(_rnd.nextInt(_chars.length)),
+      ),
+    );
 
 RenderBox findAndUpdateWidgetPosition({
   required GlobalKey widgetAnchor,
@@ -19,9 +31,9 @@ RenderBox findAndUpdateWidgetPosition({
   final provider = context.read<VSNodeDataProvider>();
 
   data.widgetOffset =
-      provider.applyViewPortTransfrom(position) - data.nodeData.widgetOffset;
+      provider.applyViewPortTransfrom(position) - data.nodeData!.widgetOffset;
 
-  provider.updateOrCreateNodes([data.nodeData]);
+  provider.updateOrCreateNodes([data.nodeData!]);
 
   return renderBox;
 }