+++ lib/widgets/vs_node_output.dart
@@ -3,6 +3,7 @@ import 'package:provider/provider.dart';
 import 'package:vs_node_view/common.dart';
 import 'package:vs_node_view/data/vs_interface.dart';
 import 'package:vs_node_view/data/vs_node_data_provider.dart';
+import 'package:vs_node_view/special_nodes/vs_list_node.dart';
 import 'package:vs_node_view/special_nodes/vs_widget_node.dart';
 import 'package:vs_node_view/widgets/gradiant_line_drawer.dart';
 
@@ -32,6 +33,19 @@ class _VSNodeOutputState extends State<VSNodeOutput> {
   void initState() {
     super.initState();
 
+    updateRenderBox();
+  }
+
+  @override
+  void didUpdateWidget(covariant VSNodeOutput oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    if (widget.data.nodeData is VSListNode) {
+      updateRenderBox();
+    }
+  }
+
+  void updateRenderBox() {
     WidgetsBinding.instance.addPostFrameCallback((_) {
       renderBox = findAndUpdateWidgetPosition(
         widgetAnchor: _anchor,